refactor out common code
